feat: use workflow store in jobEditor via hook
